More Refactor
* Refactor To improve code quality (codeclimate.com issues)
